SodaBible Update on 1/7/16
Hello! I just updated the Welcome.js file for SodaBible. I gave it a
better description (though not by much haha), links to my company
(Soderminal), and I put a Thank You message in there as well. I will be
working on this project more (hopefully) in the future, so stay tuned
for that! Thank You for reading this commit and God Bless!